requirements.txt
asgiref==3.8.1
boto3==1.35.37
botocore==1.35.37
Django==5.1.2
django-cors-headers==4.4.0
django-storages==1.14.4
djangorestframework==3.15.2
drf-yasg==1.21.7
inflection==0.5.1
jmespath==1.0.1
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
s3transfer==0.10.3
setuptools==75.1.0
six==1.16.0
sqlparse==0.5.1
uritemplate==4.1.1
urllib3==2.2.3
djangorestframework-simplejwt~=5.3.1
celery~=5.4.0